third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/scripting-1/the-template-element/additions-to-the-steps-to-clone-a-node/templates-copy-document-owner-expected.txt

CONSOLE ERROR: line 114: Uncaught ReferenceError: testInIFrame is not defined
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught ReferenceError: testInIFrame is not defined
Harness: the test ran to completion.
